moved event listener to events hook
