+++ sandbox/src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import { web3ProviderConfig } from './providerConfig';
+import { web3ProviderConfig } from './examples/web3ProviderConfig';
 import { Web3Provider } from '@decent-org/wallet-provider';
 import { ToastContainer } from 'react-toastify';
 import "react-toastify/dist/ReactToastify.css";
